backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
python-dotenv==1.0.0
vertexai==0.0.1
google-cloud-aiplatform==1.36.0
tenacity==8.2.3
pillow==10.0.1
